Improve encapsulation of `moxie_ui` widgets/elems (#130)
The goal here is to encapsulate ui elements and widgets into their own
modules so private `const`s, `fn`s, `struct`s stays private without
leaking into unnecessary places.

This PR also deletes most of the marker components where it's become
useless because of the use of `moxie_ui_kernel` bindings & watchers.
