+++ editor/moxie/src/lib.rs
@@ -10,25 +10,18 @@
 // Inherent to Bevy ECS: systems take many params and query tuples.
 #![allow(clippy::type_complexity, clippy::too_many_arguments)]
 
-mod hierarchy;
 mod icons;
 mod playback;
-mod scene;
+mod ui;
 mod view;
 
 use core::time::Duration;
 
-use bevy::feathers::FeathersPlugins;
-use bevy::feathers::dark_theme::create_dark_theme;
-use bevy::feathers::theme::UiTheme;
 use bevy::prelude::*;
 use bevy::settings::{
     ReflectSettingsGroup, SettingsGroup, SettingsPlugin,
 };
 use bevy_motiongfx::prelude::TimelineId;
-use moxie_ui::dock::DockPlugin;
-use moxie_ui::inspector::InspectAppExt;
-use moxie_ui::reactive::{KernelPlugin, KernelSet};
 
 /// Plugin that renders a timeline editor UI for the first
 /// [`Timeline`](bevy_motiongfx::prelude::BevyTimeline).
@@ -39,37 +32,7 @@ impl Plugin for MoxiePlugin {
         app.add_plugins(SettingsPlugin::new(
             "org.voxell.motiongfx.editor",
         ))
-        .add_plugins(EditorUiPlugin);
-    }
-}
-
-/// Wires feathers theming, the editor scene, and the per-frame
-/// timeline/playback/preview systems.
-struct EditorUiPlugin;
-
-impl Plugin for EditorUiPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_plugins((
-            FeathersPlugins,
-            DockPlugin,
-            KernelPlugin::new(scene::build_editor_ui),
-        ))
-        .register_default_inspects()
-        // Seed the feathers palette (its default theme is empty).
-        .insert_resource(UiTheme(create_dark_theme()))
-        .init_resource::<EditorState>()
-        .add_systems(Startup, scene::setup_editor_ui)
-        .add_systems(
-            Update,
-            (
-                playback::play_pause_hotkey,
-                playback::stop_at_track_end,
-                view::retarget_scene_cameras,
-            )
-                .chain()
-                .before(KernelSet),
-        )
-        .add_observer(playback::on_toggle_playback);
+        .add_plugins(ui::UiPlugin);
     }
 }
 
@@ -81,15 +44,6 @@ pub(crate) const PIXELS_PER_SECOND: f32 = 160.0;
 pub(crate) fn px_for(t: Duration) -> f32 {
     t.as_secs_f32() * PIXELS_PER_SECOND
 }
-pub(crate) const PANEL_PADDING: f32 = 12.0;
-pub(crate) const NAME_PANEL_WIDTH: f32 = 140.0;
-pub(crate) const NAME_PANEL_MIN: f32 = 60.0;
-pub(crate) const NAME_PANEL_MAX: f32 = 400.0;
-pub(crate) const CONTROL_BAR_HEIGHT: f32 = 40.0;
-pub(crate) const TRACK_TOP_PADDING: f32 = 12.0;
-/// Height of one track's box, and the gap below it.
-pub(crate) const TRACK_HEIGHT: f32 = 22.0;
-pub(crate) const TRACK_GAP: f32 = 4.0;
 
 /// The offscreen texture the composition's scene cameras render into.
 /// `bevy_ui` scales this image to fit the preview area above the
