+++ editor/moxie_ui/src/lib.rs
@@ -1,10 +1,4 @@
-//! Reusable `bevy_ui` widgets for the MotionGfx editor: [`dock`]
-//! (docking engine), [`glass`] (frosted-glass material), [`inspector`]
-//! (reflect inspector), [`reactive`] (kernel adapter), and [`theme`].
-//!
-//! UI-only: no `bevy_motiongfx` or editor domain deps, so the dock
-//! system is reusable standalone. Widgets build on the headless
-//! [`bevy::ui_widgets`] behaviors and the [`bevy::feathers`] theme.
+//! Reusable UI elements and widgets for Moxie.
 
 #![allow(
     clippy::type_complexity,
@@ -12,120 +6,43 @@
     reason = "Inherent to Bevy ECS: systems take many params and query tuples."
 )]
 
-pub mod dock;
+pub mod elements;
 pub mod glass;
-pub mod inspector;
+pub mod icons;
 pub mod reactive;
 pub mod theme;
+pub mod widgets;
 
-use bevy::feathers::cursor::EntityCursor;
-use bevy::feathers::theme::{ThemeBackgroundColor, ThemedText};
-use bevy::feathers::tokens;
+use bevy::feathers::FeathersPlugins;
+use bevy::feathers::dark_theme::create_dark_theme;
+use bevy::feathers::theme::UiTheme;
 use bevy::prelude::*;
-use bevy::ui_widgets::ControlOrientation;
-use bevy::window::SystemCursorIcon;
 
-pub const PLAYHEAD_COLOR: Color = Color::srgb(0.95, 0.30, 0.35);
+use glass::GlassPlugin;
+use reactive::KernelPlugin;
+use widgets::dock::DockPlugin;
+use widgets::inspector::InspectAppExt;
 
-/// A theme-inheriting text label.
-pub fn label<M: Component + Default + Unpin + Clone>(
-    text: &str,
-) -> impl Scene {
-    bsn! {
-        M
-        Text({text})
-        ThemedText
-        TextFont {
-            font_size: FontSize::Px(13.0)
-        }
-
-    }
-}
-
-/// The playhead line, positioned by the editor's playhead system.
-pub fn playhead_line(left: f32) -> impl Scene {
-    bsn! {
-        Node {
-            position_type: PositionType::Absolute,
-            top: Val::Px(0.0),
-            bottom: Val::Px(0.0),
-            left: Val::Px(left),
-            width: Val::Px(2.0),
-        }
-        ZIndex(10)
-        BackgroundColor(PLAYHEAD_COLOR)
-    }
-}
-
-/// The scrubbable timeline track: a plain node sized to the track's
-/// duration (`PIXELS_PER_SECOND` per second), so a clip at time `t`
-/// sits at `t * PIXELS_PER_SECOND` from its left edge.
+/// Everything a consumer needs to render a moxie UI: feathers theming,
+/// the dock engine (which pulls in [`glass::GlassPlugin`]), the
+/// default reflect-inspector widgets, and the kernel.
 ///
-/// Scrubbing is driven by pointer observers on this node (see
-/// the editor's track pointer observers) rather than a
-/// headless `Slider`: a scrub can only *begin* from a press that
-/// actually lands inside the track, so it can't be started from
-/// elsewhere in the window.
-pub fn timeline_track(width: f32) -> impl Scene {
-    bsn! {
-        Node {
-            position_type: PositionType::Relative,
-            width: Val::Px(width),
-            min_width: Val::Px(width),
-            height: Val::Percent(100.0),
-        }
-    }
-}
+/// Doesn't build a root itself — spawn one and call
+/// [`reactive::build_root`] wherever the app does its own `Startup`
+/// setup.
+#[derive(Default)]
+pub struct MoxieUiPlugin;
 
-pub const DIVIDER_WIDTH: f32 = 6.0;
-
-#[derive(SceneComponent, Default, Clone)]
-#[scene(DividerProps)]
-pub struct Divider;
-
-pub struct DividerProps {
-    pub thickness: Val,
-    pub orientation: ControlOrientation,
-}
-
-impl Default for DividerProps {
-    fn default() -> Self {
-        Self {
-            thickness: Val::Px(DIVIDER_WIDTH),
-            orientation: ControlOrientation::Horizontal,
-        }
-    }
-}
-
-impl Divider {
-    pub fn scene(
-        DividerProps {
-            thickness,
-            orientation,
-        }: DividerProps,
-    ) -> impl Scene {
-        let (height, width, cursor_icon) = match orientation {
-            ControlOrientation::Horizontal => (
-                thickness,
-                Val::Percent(100.0),
-                SystemCursorIcon::NsResize,
-            ),
-            ControlOrientation::Vertical => (
-                Val::Percent(100.0),
-                thickness,
-                SystemCursorIcon::EwResize,
-            ),
-        };
-        bsn! {
-            Divider
-            Node {
-                width,
-                height,
-                flex_shrink: 0.0,
-            }
-            ThemeBackgroundColor(tokens::PANE_HEADER_DIVIDER)
-            EntityCursor::System(cursor_icon)
-
-        }
+impl Plugin for MoxieUiPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugins((
+            FeathersPlugins,
+            DockPlugin,
+            KernelPlugin,
+            GlassPlugin,
+        ))
+        .register_default_inspects()
+        // Seed the feathers palette (its default theme is empty).
+        .insert_resource(UiTheme(create_dark_theme()));
     }
 }
