+++ editor/moxie_ui/src/widgets/dock/registry.rs
