+++ editor/moxie_ui/src/widgets/dock/split.rs
