+++ editor/moxie_ui/src/widgets/inspector.rs
